refactor: rename util modules to concept names
`poll_util` and `time_util` named the meta-category ("util") instead of
the concept they embody. Rename them to what a reader would search for:

- `time_util` -> `cadence` (anchor-phase cadence arithmetic)
- `poll_util` -> `noop_waker` (no-op-waker poll context)

Both modules are `pub(crate)`, so this is a pure internal rename with no
public-API or semver impact. Function names were already conceptual and
are left unchanged.
