+++ src/lib.rs
@@ -92,9 +92,10 @@
 //! Enables `subscription::http` with Query and Mutation support.
 
 pub(crate) mod application;
+pub(crate) mod cadence;
 pub mod command;
+pub(crate) mod noop_waker;
 pub(crate) mod panic;
-pub(crate) mod poll_util;
 pub mod prelude;
 pub(crate) mod runtime;
 mod structural_key;
@@ -104,7 +105,6 @@ pub mod subscription;
 // app never names the type (RFC 0008 §3.3; docs/api-guidelines.md "Prelude
 // Membership").
 pub mod testing;
-pub(crate) mod time_util;
 
 // Re-export commonly used types
 pub use application::Application;
